Update 01.requirements.md
Hi there i was a technical writer in germany and i love the simplicity of your brainchild htmly. Now there are "please improve documentation"-signs all over the place. So i thought that i should support this marvellous piece oft software. I here just changed the first paragraph of the first documnetaion and would be interested, which part of the documentation is currently in most need of improvement. please return your thoughts to info@clauswilcke.com
all the best
Claus